\name{funArray}
\alias{funArray}
\title{
  Create an array of functions
}
\description{
  Creates a array of functions.
}
\usage{
funArray(\ldots)
}
\arguments{
  \item{\ldots}{
    passed to \code{\link{array}}. see arguments to \code{\link{array}}
  }
}
\author{
  Charles Dupont \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr  
}
\seealso{
  \code{\link{funVector}}, \code{\link{funMatrix}}, \code{\link{array}}
}
\keyword{ classes }
