\name{funMatrix}
\alias{funMatrix}
\title{
  Create a matrix of functions
}
\description{
  Creates a matrix of functions.
}
\usage{
funMatrix(\ldots)
}
\arguments{
  \item{\ldots}{
    passed to \code{\link{matrix}}. see arguments to \code{\link{matrix}}
  }
}
\author{
  Charles Dupont \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr  
}
\seealso{
  \code{\link{funVector}}, \code{\link{funArray}}, \code{\link{matrix}}
}
\keyword{ classes }
