\name{funVector}
\alias{funVector}
\title{
  Create a vector of functions
}
\description{
  Creates a vector of functions.
}
\usage{
funVector(length = 0)
}
\arguments{
  \item{length}{
    integer; length of vector.
  }
}
\author{
  Charles Dupont \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr  
}
\seealso{
  \code{\link{funMatrix}}, \code{\link{funArray}}, \code{\link{vector}}
}
\keyword{ classes }
