\name{calc.v}
\alias{calc.v}
\title{
  Calculates the v matrix.
}
\description{
  Calculates the v matrix.
}
\usage{
calc.v(event, time)
}
\arguments{
  \item{event}{
    logical vector indicating whether and event has happend.
  }
  \item{time}{
    vector; time until event or observation halted.
  }
}
\value{
  returns a matrix.
}
\references{
 ~put references to the literature/web site here ~
}
\author{
  Bryan E. Shepherd \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr

  
  Charles Dupont \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{design}
