\name{sensitivityPStrat-package}
\alias{sensitivityPStrat-package}
\alias{sensitivityPStrat}
\docType{package}
\title{
  Principal Stratification Sensitivity Analysis Functions
}
\description{
  This package provides functions to perform principal stratification
  sensitivity analyses on datasets.
}
\details{
  sensitivityPStrat provides functions for performing principal
  stratification sensitivity analysis. It implements functions as
  described in: 

  Gilbert PB, Bosch RJ, and Hudgens MG (2003), "Sensitivity Analysis for
  the Assessment of Causal Vaccine Effects of Viral Load in HIV Vaccine
  Trials," Biometrics 59, 531-541.
  
  Hudgens MG, Hoering A, and Self SG (2003), "On the Analysis of Viral
  Load Endpoints in HIV Vaccien Trials," Statistics in Medicine 22,
  2281-2298.

  Jemiai Y (2005), "Semiparametric Methods for Inferring Treatment
  Effects on Outcomes Defined Only if a Post-Randomization Event
  Occurs," unpublished doctoral dissertation under the supervision of
  A. Rotnitzky, Harvard School of Public Health, Dept. of Biostatistics.
  
  Shepherd BE, Gilbert PB, Lumley T (2007), "Sensitivity analyses
  comparing time-to-event outcomes existing only in a subset selected
  postrandomization," Journal of the American Statistical Association
  102, 573-582.
  
  Shepherd BE, Gilbert PB, and Dupont CT, "Sensitivity analyses
  comparing time-to-event outcomes only existing in a subset selected
  postrandomization and relaxing monotonicity," (submitted paper).
}
\author{
  Bryan E. Shepherd \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr

  
  Charles Dupont \cr
  Department of Biostatistics\cr
  Vanderbilt University\cr
  \email{charles.dupont@vanderbilt.edu}

  Maintainer: Charles Dupont
}
\references{
  Gilbert PB, Bosch RJ, and Hudgens MG (2003), "Sensitivity Analysis for
  the Assessment of Causal Vaccine Effects of Viral Load in HIV Vaccine
  Trials," Biometrics 59, 531-541.
  
  Hudgens MG, Hoering A, and Self SG (2003), "On the Analysis of Viral
  Load Endpoints in HIV Vaccien Trials," Statistics in Medicine 22,
  2281-2298.

  Jemiai Y (2005), "Semiparametric Methods for Inferring Treatment
  Effects on Outcomes Defined Only if a Post-Randomization Event
  Occurs," unpublished doctoral dissertation under the supervision of
  A. Rotnitzky, Harvard School of Public Health, Dept. of Biostatistics.
  
  Shepherd BE, Gilbert PB, Lumley T (2007), "Sensitivity analyses
  comparing time-to-event outcomes existing only in a subset selected
  postrandomization," Journal of the American Statistical Association
  102, 573-582.
  
  Shepherd BE, Gilbert PB, and Dupont CT, "Sensitivity analyses
  comparing time-to-event outcomes only existing in a subset selected
  postrandomization and relaxing monotonicity," (submitted paper).
}
\keyword{ package }
\seealso{
  \code{\link[survival]{Surv}}
}
